kernel: Allow building dtbo.img
taimen uses the dtbo.img as an DTB Overlay in order to use
the same defconfig for both, taimen and wahoo. The dtbo.img
is built with "make dtbo.img". Allow automatically building
it when setting TARGET_NEEDS_DTBOIMAGE to true, or by manually
calling make dtboimage.

Thanks to luca020400 for help with the kernel.mk logic

Change-Id: I3bdd961535b6e60c70ad35ae4653c95a04137221
Signed-off-by: mydongistiny <jaysonedson@gmail.com>